backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
python-multipart>=0.0.9
pydantic>=2
pydantic-settings>=2
numpy>=1.24
pillow>=10
opencv-python-headless>=4.8
requests>=2.31
# ultralytics is installed on the cloud conda env already
# ultralytics>=8.3.237
